Fiddle with lockfile destructor
There are more subleties here than I originally thought. I think the general idea
is that we want `fs::rename` XOR `fs::remove_file` to succeed and therefore clean
up the lock file. The original implementation was incorrect: if we rename in the
destructor, it may be called during error handling while writing data, leaving it
in an invalid state. So we need an explicit `commit` method to signify that we're
ready to rename.

We do want to remove the lockfile during the destructor, so it's deleted in case
of error, but we do **NOT** want to remove the lockfile after renaming, as another
process may have acquired the lock. I'm not sure how to work around this other than
(ab)using [`std::mem::forget`][fg] to avoid running the destructor after renaming.

[fg]: https://doc.rust-lang.org/std/mem/fn.forget.html